added container.waitForBlock
