+++ packages/testcontainers/src/chains/main_net_container.ts
@@ -1,10 +1,10 @@
 import { DockerOptions } from 'dockerode'
-import { DeFiDContainer } from './container'
+import { DeFiDContainer } from './defid_container'
 
 export class MainNetContainer extends DeFiDContainer {
   /**
-   * @param {string} image docker image name
-   * @param {DockerOptions} options
+   * @param {string} [image=DeFiDContainer.image] docker image name
+   * @param {DockerOptions} [options]
    */
   constructor (image: string = DeFiDContainer.image, options?: DockerOptions) {
     super('mainnet', image, options)
