+++ packages/testcontainers/src/chains/test_net_container.ts
@@ -1,10 +1,10 @@
 import { DockerOptions } from 'dockerode'
-import { DeFiDContainer, StartOptions } from './container'
+import { DeFiDContainer, StartOptions } from './defid_container'
 
 export class TestNetContainer extends DeFiDContainer {
   /**
-   * @param {string} image docker image name
-   * @param {DockerOptions} options
+   * @param {string} [image=DeFiDContainer.image] docker image name
+   * @param {DockerOptions} [options]
    */
   constructor (image: string = DeFiDContainer.image, options?: DockerOptions) {
     super('testnet', image, options)
