+++ packages/testcontainers/src/index.ts
@@ -2,7 +2,7 @@ export { DockerOptions } from 'dockerode'
 
 export * from './testkeys'
 
-export * from './chains/container'
+export * from './chains/defid_container'
 export * from './chains/main_net_container'
 export * from './chains/test_net_container'
 export * from './chains/reg_test_container/index'
